Support X-Accel-Redirect in nginx >=1.5.9
Newer versions of nginx require a properly encoded URI in the X-Accel-Redirected response. According to research done by people on the django-sendfile issue tracker, the specific version number seems to be 1.5.9.

The ``PRIVATE_STORAGE_NGINX_VERSION`` setting is used to determine the proper encoding (defaults to current behaviour of not quoting).

https://github.com/adnrs96/django-sendfile/commit/4488e520f04661136e88be38281e789ae50a260c