Use 'six' API directly
As Django 3 drops compatibility with Python 2.7 and removes a number of
private APIs, importing `django.utils.six` is no longer possible, but
`string_types` can be imported from `six` directly.

Note that this change is compatible with Python 2.7, meaning
`django-private-storage` would stay compatible with Python 2.7 if this
commit is merged.