Use 'lru_cache' from 'functools'
As Django 3 drops compatibility with Python 2.7 and removes a number of
private APIs, importing `django.utils.lru_cache` is no longer possible, but
`lru_cache` can be imported from `functools` instead.

Note that while this change is compatible with all *Django* versions
down to 1.8, it breaks compatibility of `django-private-storage` with
Python 2.7 as `lru_cache` was not introduced to `functools` until Python
3.2.

Merging both #42 and this PR would make `django-private-storage` work
with Django 3 (AFAICT, at least the unit tests succeed), but at the
price of losing compatibility with Python 2.7.